Fixed and updated nuget packages
Fixed project references to use VSSDK nuget packages
