+++ .gitignore
@@ -1,7 +1,3 @@
-/.manifest
-/.pecrypt-link
-/.pecrypt-link-updated
-
 bin
 obj
 packages
@@ -9,19 +5,21 @@ out
 temp
 /lib
 .vs
-*.boltdata
 
 *.suo
 *.sdf
 *.userprefs
 *.user
 *.nupkg
-*.t4properties
-*.DS_Store
 *.sln.ide
 *.metaproj
 *.tmp
 *.cache
+project.lock.json
+*.nuget.props
+*.nuget.targets
+
+
 
 log.txt
 NuGet.exe
