--- src/NuGet.Packaging.VisualStudio/packages.config
@@ -1,28 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Clarius.VisualStudio" version="1.3.7" targetFramework="net46" />
-  <package id="Fody" version="1.29.2" targetFramework="net46" developmentDependency="true" />
-  <package id="Microsoft.Composition" version="1.0.27" targetFramework="net46" />
-  <package id="Microsoft.Tpl.Dataflow" version="4.5.24" targetFramework="net46" />
-  <package id="Microsoft.VisualStudio.Composition" version="14.2.19-pre" targetFramework="net46" />
-  <package id="Microsoft.VisualStudio.ImageCatalog" version="14.0.23205" targetFramework="net46" />
-  <package id="Microsoft.VisualStudio.Imaging.Interop.14.0.DesignTime" version="14.0.23205" targetFramework="net46" />
-  <package id="Microsoft.VisualStudio.ProjectSystem" version="14.1.127-pre" targetFramework="net46" />
-  <package id="Microsoft.VisualStudio.ProjectSystem.Analyzers" version="14.1.127-pre" targetFramework="net46" developmentDependency="true" />
-  <package id="Microsoft.VisualStudio.ProjectSystem.SDK" version="14.1.127-pre" targetFramework="net46" />
-  <package id="Microsoft.VisualStudio.ProjectSystem.SDK.Tools" version="14.1.127-pre" targetFramework="net46" developmentDependency="true" />
-  <package id="Microsoft.VisualStudio.SDK.VsixSuppression" version="14.1.15" targetFramework="net46" />
-  <package id="Microsoft.VisualStudio.Threading" version="14.1.131" targetFramework="net46" />
-  <package id="Microsoft.VisualStudio.Threading.Analyzers" version="14.1.131-pre" targetFramework="net46" developmentDependency="true" />
-  <package id="Microsoft.VisualStudio.Validation" version="14.1.111" targetFramework="net46" />
-  <package id="PropertyChanged.Fody" version="1.51.3" targetFramework="net46" developmentDependency="true" />
-  <package id="System.Collections" version="4.0.0" targetFramework="net46" />
-  <package id="System.Collections.Immutable" version="1.1.37" targetFramework="net46" />
-  <package id="System.Diagnostics.Debug" version="4.0.0" targetFramework="net46" />
-  <package id="System.Globalization" version="4.0.0" targetFramework="net46" />
-  <package id="System.Linq" version="4.0.0" targetFramework="net46" />
-  <package id="System.Resources.ResourceManager" version="4.0.0" targetFramework="net46" />
-  <package id="System.Runtime" version="4.0.0" targetFramework="net46" />
-  <package id="System.Runtime.Extensions" version="4.0.0" targetFramework="net46" />
-  <package id="System.Threading" version="4.0.0" targetFramework="net46" />
-</packages>
